Flytter state til folder for endringer
